requirements.txt
faiss-cpu
rank_bm25
streamlit
rouge-score
python-dotenv
langchain==0.1.16 
langchain-core==0.1.53
langchain-community==0.0.34
langchain-openai==0.0.8